4.0.1: Handle upstream type constraints
Previously, if ElementA had a field with a type constraint:
`1..1 Foo is Bar`
Then if ElementB was based on ElementA and tried to constrain `Bar` further:
`     Bar is ABC#123`
... the expansion would fail to recognize that ElementB's Bar constraint is actually a further constraint on Element A's Foo .  Now that is fixed.

In addition, invalid overrides (such as picking an invalid option from a choice) used to result in a null field which threw an error before the root cause error could be printed out.  This is now fixed so that such errors don't result in a null field -- instead they default the field back to its non-overridden state.

Bumped up to version 4.0.1 so we can publish the change.