Expand Choice Value Mappings
When a mapping uses the special "Value" word in its path, and that value is actually a choice, we will now expand that into a new mapping rule per choice option.

This is the only way to support mapping choice values as the time, since we don't support passing the special "Value" path down to the fhir exporter.