Expand all mapping rules in context of element
Prior to this commit, any inherited mapping rules were expanded in the
context of their based on classes.  This means that paths like `Value`
were expanded to the based on value, not the element's value (if it
was constrained) -- and fields that were constrained to a sub-type in
the element might be expanded based on the parent (not the sub-type).

This commit fixes that by collecting all of the inherited rules before
they are expanded, and then expands them in the context of the element.