Allow constraint fields on references
The grammar doesn't allow you to easily constrain a ref field with another ref.

For example, if `A` has field `ref(B)`, the grammar doesn't allow `SubA` to say:
`ref(B) is type SubB`.
Instead, you have to say:
`B is type SubB`
... but then expand complains because you override a (`ref(B)`) with a non-ref (`B`).

This commit allows the latter so authors don't have to repeate the `ref` in sub-classes.